requirements.txt
PyQt5==5.15.6
PyQtWebEngine==5.15.6
Pillow==9.0.1
requests==2.27.1 